API

QAOALandscapes.QAOA — Type

QAOA(g::T; applySymmetries=true) where T <: AbstractGraph

Constructors for the QAOA object.

QAOALandscapes.HxDiag — Method

HxDiag(g::T) where T<: AbstractGraph

Construct the mixing Hamiltonian. If the cost Hamiltonian is invariant under the parity operator $\prod_{i=1}^N \sigma^x_i$ it is better to work in the +1 parity sector of the Hilbert space since this is more efficient. In practice, if the system size is $N$, the corresponding Hamiltonianwould be a vector of size $2^{N-1}$.

QAOALandscapes.HxDiagSymmetric — Method

HxDiagSymmetric(T::Type{<:Real}, g::S) where S <: AbstractGraph

Construct the mixing Hamiltonian in the positive (+1) parity sector of the Hilbert space. This means that if the system size is N, then HxDiagSymmetric would be a vector of size $2^{N-1}$. This construction, only makes sense if the cost/problem Hamiltonian $H_C$ is invariant under the action of the parity operator, that is

\[ [H_C, \prod_{i=1}^N \sigma^x_i] = 0\]

QAOALandscapes.HzzDiag — Method

HzzDiag(g::T) where T <: AbstractGraph

Construct the cost Hamiltonian. If the cost Hamiltonian is invariant under the parity operator $\prod_{i=1}^N \sigma^x_i$ it is better to work in the +1 parity sector of the Hilbert space since this is more efficient. In practice, if the system size is $N$, the corresponding Hamiltonian would be a vector of size $2^{N-1}$. This function instead returs a vector of size $2^N$.

QAOALandscapes.HzzDiagSymmetric — Method

HzzDiagSymmetric(g::T) where T <: AbstractGraph
HzzDiagSymmetric(edge::T) where T <: AbstractEdge

Construct the cost Hamiltonian in the positive (+1) parity sector of the Hilbert space. This means that if the system size is N, then HzzDiagSymmetric would be a vector of size $2^{N-1}$. This construction, only makes sense if the cost/problem Hamiltonian $H_C$ is invariant under the action of the parity operator, that is

\[ [H_C, \prod_{i=1}^N \sigma^x_i] = 0\]

Similarly, if the input is an edge then it returns the corresponding $ZZ$ operator.

QAOALandscapes.computationalBasisWeights — Method

computationalBasisWeights(ψ, equivClasses)

Computes the computational basis weights for a given state vector ψ according to the provided equivalence classes, by summing up the squared magnitudes of the elements with the same equivalence class. Here ψ and equivClasses are ment to live in the same Hilbert space basis.

Arguments

ψ: The state vector.

equivClasses: A Vector of Vectors, where each inner Vector contains the indices of the elements belonging to the same equivalence class.

Returns

basis_weights::Vector{Float64}: A Vector containing the summed squared magnitudes of ψ for each equivalence class.

QAOALandscapes.fourierOptimize — Method

fourierOptimize(qaoa::QAOA, Γ0::Vector{Float64}, pmax::Int)

Starting from a local minima Γ0 at $p=1$ it performs the Fourier optimization strategy until the circuit depth pmax is reached. By default the BFGS optimizer is used.

Arguments

qaoa::QAOA: QAOA object

Γ0::Vector{Float64}: Vector correponding to the local minimum from which we will construct the particular TS and then roll down from.

Return

result:Dict. Dictionary with keys being keys \in [1, pmax] and values being a Tuple{Float64, Vector{Float64}} of cost function value and corresponding parameter.

QAOALandscapes.gad — Method

gad(qaoa::QAOA, init_point::Vector{T}; niter = 500, η=0.01, tol=1e-5) where T<:Real

Perform the Gentlest Ascent Dynamics (GAD) optimization algorithm on a QAOA problem with the goal to find index-1 saddle points.

Arguments

qaoa::QAOA: a QAOA problem instance.

init_point::Vector{T}: the initial point in parameter space, where T is a subtype of Real.

niter::Int=500: the maximum number of iterations to perform (optional, default is 500).

η::Float64=0.01: the step size for the GAD algorithm (optional, default is 0.01).

tol::Float64=1e-5: the tolerance for the gradient norm. If the norm falls below this value, the algorithm stops (optional, default is 1e-5).

Returns

point_history: history of points during the iterations.

energ_history: history of energy values during the iterations.

grad_history: history of gradient norms during the iterations.

Usage

```julia pointhistory, energhistory, gradhistory = gad(qaoa, initpoint, niter=500, η=0.01, tol=1e-5)

QAOALandscapes.generalClassicalHamiltonian — Method

generalClassicalHamiltonian(interaction_dict::Dict{Vector{Int64}, Float64})

This function computes the classical Hamiltonian for a general $p$-spin Hamiltonian, that is

\[ H_Z = \sum_{i \in S} J_i \prod_{\alpha \in i} \sigma^z_{i_{\alpha}}\]

Above, $S$ is the set of interaction terms which is passed as an argument in the for of a dictionary with keys being the spins participating in a given interaction and values given by the weights of such interaction term.

Arguments

interaction_dict: a dictionary where each key is a vector of integers representing an interaction, and each value is the weight of that interaction.

Returns

hamiltonian::Vector{Float64}: The general $p$ spin Hamiltonian.

QAOALandscapes.geometricTensor — Method

geometricTensor(qaoa::QAOA, params::Vector{T}, ψ0::AbstractVector{Complex{T}}) where T<: Real

Compute the geometricTensor of the QAOA cost function using adjoint (a reverse-mode) differentiation. We implement the algorithm proposed in this reference

QAOALandscapes.getEquivalentClasses — Method

getEquivalentClasses(vec::Vector{T}; sigdigits = 5) where T <: Real

Computes the equivalence classes of the elements of the input vector vec The elements are rounded to a certain number of significant digits (default is 5) to group the states with approximately equal values.

Arguments

vec::Vector{T<:Real}

sigdigits=5: Significant digits to which energies are rounded.

Returns

data_states::Dict{Float64, Vector{Int}}: Keys are unique elements (rounded) and values corresponds to the index of elements with the same key.

QAOALandscapes.getHessianIndex — Method

getHessianIndex(qaoa::QAOA, Γ::AbstractVector{T}; checks=true, tol=1e-6) where T<:Real

Calculate the Hessian index of a stationary (it checks the gradient norm) point of the QAOA energy function

Arguments

qaoa: a QAOA object.

Γ: a vector of parameters.

Keyword Arguments

checks=true: a boolean to decide whether to check if the gradient of the cost function is smaller than a certain tolerance.

tol=1e-6: a tolerance level for the gradient of the cost function.

Output

Returns the Hessian index, i.e., the number of negative eigenvalues of the Hessian matrix.

Notes

The function first calculates the gradient of the cost function for the given qaoa and Γ. If checks=true, it asserts that the norm of this gradient is less than tol. It then calculates the Hessian matrix and its eigenvalues, and returns the count of eigenvalues less than zero.

QAOALandscapes.getInitialParameter — Method

getInitialParameter(qaoa::QAOA; spacing = 0.01, gradTol = 1e-6)

Given a QAOA object it performs a grid search on a region of the two dimensional space spanned by $\{ \gamma_1, \beta_1\}$ The $\beta_1$ component is in the interval $[-\pi/4, \pi/4]$ while the $\gamma_1$ part is in the $(0, \pi/4]$ for 3RRG or $(0, \pi/2]$ for dRRG (with $d\neq 3$).

We then launch the QAOA optimization procedure from the point in the 2-dimensional grid with the smallest cost function value.

Returns

3-Tuple containing: 1.) the cost function grid, 2.) the optimal parameter, and 3.) the optimal energy

QAOALandscapes.getNegativeHessianEigval — Method

getNegativeHessianEigval(qaoa::QAOA, Γmin::Vector{Float64}, ig::Int; tsType="symmetric")

Computes the approximation to the minimum (negative) eigenvalue of the Hessian at the TS obtained by padding with zeros the local minimum Γmin. The transition state is completely specified by the index of the γ component ig, and the type of transition states ("symmetric" or "non_symmetric"). The cost of obtaining this approximate eigenvalue is basically the cost of computing two matrix elements of a Hessian.

QAOALandscapes.getNegativeHessianEigvec — Method

getNegativeHessianEigvec(qaoa::QAOA, Γmin::Vector{Float64}, ig::Int; tsType="symmetric", doChecks=false)

Computes the approximate eigenvalue and index-1 eigenvector of the Hessian at the transition state obtained from the local minimum Γmin. It is completely specified by the parameters iγ and tsType="symmetric". If the optional parameter doChecks=false is set to true, then the function also returns the relative error in estimating the true eigenvalue as well as the inner product between the approximate and true eigenvector

Arguments

qaoa::QAOA: QAOA object

Γmin::Vector{Float64}: The vector corresponding to a local minimum of QAOAₚ.

ig::Int: Index of the γ component where the zeros are inserted.

tsType="symmetric": In this case, the index of the β component is equal to ig. Otherwise, the β index is ig-1.

Optional arguments

doChecks=false: In this case the function returns a dictionary with keys eigvec_approx and eigval_approx. If set to true it has additional keys => change_basis, eigvec_fidelity and eigval_error

Returns

result::Dict Dictionary with the following keys: eigvec_approx, eigval_approx. If doChecks=true the following additional keys are available: change_basis, eigvec_fidelity and eigval_error

QAOALandscapes.getQAOAState — Method

getQAOAState(q::QAOA, Γ::AbstractVector{T}) where T <: Real

Construct the QAOA state. More specifically, it returns the state:

\[ |\Gamma^p \rangle = U(\Gamma^p) |+\rangle\]

with

\[ U(\Gamma^p) = \prod_{l=1}^p e^{-i H_{B} \beta_{2l}} e^{-i H_{C} \gamma_{2l-1}}\]

and $H_B, H_C$ corresponding to the mixing and cost Hamiltonian correspondingly.

QAOALandscapes.getStateProjection — Method

getStateProjection(qaoa::QAOA, params, stateIndex::Vector{Int64})

Calculates the projection of the QAOA state onto the state subspace determined by gsIndex. It also returns the corresponding orthogonal complement. The QAOA state is determined by the given parameters params.

Arguments

qaoa::QAOA: QAOA object.

params: Parameters determining the QAOA state.

stateIndex::Vector{Int64}: Indices of the ground state components.

Returns

normState: Norm of the projection of the QAOA state onto the state subspace given by stateIndex.

normState_perp: Norm of the projection of the QAOA state onto the orthogonal complement of the specificied state subspace.

ψIndex: Normalized projection of the QAOA state onto the state subspace.

ψIndex_perp: Normalized projection of the QAOA state onto the orthogonal complement of the state subspace.

QAOALandscapes.getStationaryPoints — Method

getStationaryPoints(
    qaoa::QAOA, 
    p::Integer,
    grid_of_points;
    printout = false, 
    threaded = false,
    method = Optim.BFGS(linesearch = Optim.BackTracking(order=3))
)

Computes the stationary points of the QAOA cost function given an initial grid of points. The finer the grid the more points we should find (until saturation is reached)

The optimization is performed using the provided method (BFGS with backtracking linesearch by default). The function can operate either in single-threaded or multi-threaded mode.

Arguments

qaoa::QAOA: QAOA problem instance.

p::Integer: Integer value related to the problem dimensionality.

grid_of_points: Initial points in parameter space from which the optimization begins.

printout::Bool (optional): If true, prints the final cost function value and gradient norm. Default is false.

threaded::Bool (optional): If true, uses multi-threaded mode. Default is false.

method::Optim.AbstractOptimizer (optional): Optimizer to use. Default is BFGS with backtracking linesearch of order 3.

Returns

Tuple: Returns a tuple containing the final energies and the corresponding parameters at the stationary points.

QAOALandscapes.gradSquaredNorm — Method

gradSquaredNorm(qaoa::QAOA, point::AbstractVector{T}) where T <: Real

Calculate the squared norm of the gradient of the QAOA cost function at the given point.

Arguments

qaoa::QAOA: QAOA problem instance.

point::AbstractVector{T}: Point in parameter space at which to compute the gradient squared norm.

Returns

Float64: Returns the squared norm of the gradient at the given point.

QAOALandscapes.hessianCostFunction — Method

hessianCostFunction(qaoa::QAOA, Γ::AbstractVector{T}) where T<:Real

Computes the cost function Hessian at the point $\Gamma$ in parameter space. The computation is done analytically since it has proven to be faster than the previous implementation using ForwardDiff.jl package

QAOALandscapes.interpInitialization — Method

interpInitialization(Γp::Vector{Float64})

Given an initial state Γp::Vector{Float64} of length 2p it creates another vector ΓInterp of size $2p+2$ with $\gamma (\beta)$ components given by the following expression

\[\gamma^i_{p+1} = \frac{i-1}{p} \gamma^{i-1}_{p} + \frac{p-i+1}{p}\gamma^{i}_{p}\]

and analogously for the $\beta$ components.

QAOALandscapes.interpOptimize — Method

interpOptimize(qaoa::QAOA, Γ0::Vector{Float64}, pmax::Int; method=Optim.BFGS(linesearch = Optim.BackTracking(order=3)))

Starting from a local minima Γ0 at $p=1$ it performs the Interp optimization strategy until the circuit depth pmax is reached. By default the BFGS optimizer is used.

Arguments

qaoa::QAOA: QAOA object

Γ0::Vector{Float64}: Vector correponding to the local minimum from which we will construct the particular TS and then roll down from.

Optional

method=Optim.BFGS(linesearch = Optim.BackTracking(order=3)): Default optimizer and linesearch choice. For more available choices see Optim.jl

Return

result:Dict. Dictionary with keys being keys \in [1, pmax] and values being a Tuple{Float64, Vector{Float64}} of cost function value and corresponding parameter.

QAOALandscapes.optimizeGradSquaredNorm — Method

optimizeGradSquaredNorm(qaoa::QAOA, point::AbstractVector{T}; printout=false, 
    method=Optim.BFGS(linesearch = Optim.BackTracking(order=3))) where T <: Real

Optimizes the squared norm of the gradient of the QAOA cost function given the initial parameter point.

Arguments

qaoa::QAOA: QAOA problem instance.

point::AbstractVector{T}: Initial point in parameter space to start the optimization.

printout::Bool (optional): If true, prints the final cost function value and gradient norm. Default is false.

method::Optim.AbstractOptimizer (optional): Optimizer to use. Default is BFGS with backtracking linesearch of order 3.

Returns

Tuple: Returns a tuple containing the optimized parameters and the minimum cost.

QAOALandscapes.optimizeParameters — Method

optimizeParameters(::Val{:Fourier}, qaoa::QAOA, Γ0::Vector{Float64}; printout=false)

Perform optimization of the QAOA using the gradient descent algorithm with the BFGS optimizer. Here we use the alternative "Fourier" initialization, where instead of optimizing the usual (γ, β) parameters we optimize their frecuency components $(u_{\gamma}, u_{\beta})$.

Arguments

Val(:BFGS): For using the BFGS. Alternatively, Val(:GD) for using the ADAM optimizer

qaoa:QAOA: QAOA object

Γ0::Vector{Float64}: Initial point from where the optimization starts

Keyword arguments

printout=false: Whether if we print something during the optimization or not

Returns

It returns a tuple containing the following information

parameters::Vector{Float64}: Optimal parameter obtained

cost::Float64: Value of the cost function for the optimal parameter obtained.

QAOALandscapes.optimizeParametersSlice — Method

optimizeParametersSlice(qaoa::QAOA, ΓTs::Vector{T}, u::Vector{T};
                        method=Optim.BFGS(linesearch=LineSearches.BackTracking(order=3))) where T<:Real

Optimize the parameters of the QAOA along the index-1 direction of the transition state ΓTs.

Arguments

qaoa::QAOA: QAOA instance

ΓTs::Vector{T}: Initial parameter vector. We assume that ΓTs is an index-1 saddle point and the u is the index-1 direction

u::Vector{T}: Direction along which the QAOA cost function is going to be optimized.

method: The optimization method to be used for the parameter optimization. Default: Optim.BFGS(linesearch=LineSearches.BackTracking(order=3)).

Returns

f_vals::Vector{T}: The minimum objective function values obtained from the optimization process for the negative and positive parameter shift cases.

x_vals::Vector{Vector{T}}: The corresponding parameter values that yield the minimum objective function values.

QAOALandscapes.optimizeParametersTaped — Method

optimizeParameters(qaoa::QAOA, params::AbstractVector{T};
                   method = Optim.BFGS(linesearch = Optim.BackTracking(order=3)),
                   printout::Bool = false) where T<:Real

Optimizes the QAOA parameters using the specified optimization method and linesearch algorithm.

Arguments

qaoa::QAOA: A QAOA instance representing the problem to be solved.

params::AbstractVector{T}: The initial guess for the optimization parameters, where T is a subtype of Real.

Keyword Arguments

method: The optimization method to be used (default: Optim.BFGS(linesearch = Optim.BackTracking(order=3))).

printout::Bool: Whether to print optimization progress information (default: false).

!!! Important

diffMode = :adjoint is the default but is only meant to be used for 1st and quasi 2nd order method. Use diffMode=:forward if you want to use something like Newton's methdd

Returns

A tuple containing the optimized parameters and the corresponding minimum cost function value.

Examples

result = optimizeParameters(qaoa, params, method=Optim.BFGS(linesearch=Optim.HagerZhang()), printout=true)

QAOALandscapes.permuteHessian — Method

permuteHessian(depth::Int, i::Int; tsType="symmetric")

Computes the permutation that takes the Hessian at a particular transition state into the form described in the paper Basically, the last two rows and columns of the transformed Hessian correspond to the indexes where the zeros were inserted.

Arguments

depth::Int: Circuit depth of the transition state.

i::Int: Index of the γ component at which the zero is added. If tsType='symmetric' then β=i, otherwise if tsType='non_symmetric' β=i-1.

Return

listOfIndices::Vector{Int64}: List of indices correponding to the arrangement of the new basis elements.

permMat::Matrix{Float64}: Matrix implementing the desired permutation.

QAOALandscapes.permuteHessian — Method

permuteHessian(H::AbstractArray{Float64,2}, i::Int; tsType="symmetric")

Computes the permutation that takes the Hessian at a particular transition state into the form described in the paper Basically, the last two rows and columns of the transformed Hessian correspond to the indexes where the zeros were inserted.

Arguments

H::AbstractArray{Float64,2}: Hessian at the transition state in the original basis.

i::Int: Index of the γ component at which the zero is added. If tsType='symmetric' then β=i, otherwise if tsType='non_symmetric' β=i-1.

Return

listOfIndices::Vector{Int64}: List of indices correponding to the arrangement of the new basis elements.

permMat::Matrix{Float64}: Matrix implementing the desired permutation.

HTransformed::Matrix{Float64}: Transformed Hessian at the transition state. Specifically, we have that $H_{\mathop{\rm{perm}}}=PHP^{-1}$.

QAOALandscapes.rollDownInterp — Method

rollDownInterp(qaoa::QAOA, Γmin::Vector{Float64}; method=Optim.BFGS(linesearch = Optim.BackTracking(order=3)))

Starting from a local minima we construct a new vector using the INTERP initialization from which we perform the optimization.

Arguments

qaoa::QAOA: QAOA object

Γmin::Vector{Float64}: Vector correponding to the local minimum from which we will construct the particular TS and then roll down from.

Optional

method=Optim.BFGS(linesearch = Optim.BackTracking(order=3)): Default optimizer and linesearch choice. For more available choices see Optim.jl

Return

result:Tuple. The first element corresponds to the vector corresponding to which the algorithm converged to, and the second element is correponding energy_history

QAOALandscapes.rollDownTS — Method

rollDownTS(qaoa::QAOA, Γmin::Vector{Float64}; ϵ=0.001, optim=Val(:BFGS))

Starting from a local minima we construct all transition states (a total of 2p+1 of them). From each of the transition states, we construct two new vectors

\[\Gamma^0_p = \Gamma_{\rm{TS}} + \epsilon \hat{e}_{\rm{min}}, \Gamma^0_m = \Gamma_{\rm{TS}} - \epsilon \hat{e}_{\rm{min}} \]

We then use these two vectors as initial points to carry out the optimization. Following our analytical results we are guarantee that the obtained vectors have lower energy than the initial vector Γmin

Arguments

qaoa::QAOA: QAOA object

Γmin::Vector{Float64}: Vector correponding to the local minimum from which we will construct the particular TS and then roll down from.

Return

result:Tuple. The returned paramaters are as follows => Γmin_m, Γmin_p, Emin_m, Emin_p, info_m, info_p

QAOALandscapes.rollDownfromTS — Method

rollDownTS(qaoa::QAOA, Γmin::Vector{T}, ig::Int; ϵ=0.001, tsType="symmetric")

Starting from a local minima we construct a vector corresponding to the transition state specified by ig. From there we construct two new vectors

\[\Gamma^0_p = \Gamma_{\rm{TS}} + \epsilon \hat{e}_{\rm{min}},\]

\[\Gamma^0_m = \Gamma_{\rm{TS}} - \epsilon \hat{e}_{\rm{min}} \]

We then use these two vectors as initial points to carry out the optimization. Following our analytical results we are guarantee that the obtained vectors have lower energy than the initial vector Γmin

Arguments

qaoa::QAOA: QAOA object

Γmin::Vector{Float64}: Vector correponding to the local minimum from which we will construct the particular TS and then roll down from.

ig::Int: Index of the γ component where the zeros are inserted.

tsType="symmetric": In this case, the index of the β component is equal to ig. Otherwise, the β index is ig-1.

optim=Val(:BFGS): Means that we will use the L-BFGS algorithm to perform the optimization. The other option is optim=Val{:GD}.

Return

result:Tuple. The returned paramaters are as follows => Γmin_m, Γmin_p, Emin_m, Emin_p, info_m, info_p

QAOALandscapes.spinChain — Method

spinChain(n::Int; bcond="pbc")

Constructs the graph for the classical Ising Hamiltonian on a chain with periodic boundary conditions determined by the keyword argument bcond

QAOALandscapes.toFundamentalRegion! — Method

toFundamentalRegion!(qaoa::QAOA, Γ::Vector{Float64})

Implements the symmetries of the QAOA for different graphs. For more detail see the following reference.

For an arbitrary graph, we can restrict both $\gamma$ and $\beta$ parameters to the $[-\pi/2, \pi/2]$ interval. Furthermore, $\beta$ parameters can be restricted even further to the $[-\pi/4, \pi/4]$ interval (see here) Finally, when dealing with regular graphs with odd degree \gamma paramaters can be brought to the $[-\pi/4, \pi/4]$ interval. This function modifies inplace the initial input vector $Γ$.

QAOALandscapes.transitionState — Method

transitionState(Γp::Vector{Float64})

Given an initial state Γp::Vector{Float64} of length $2p$ it creates a matrix $M_{\rm{TS}}$ of size $2p+2 \times 2p+1$. The columns of $M_{\rm{TS}}$ corresponds to the transition states associated with the initial minimum Γp. The first p+1 columns correspond to symmetric TS while the remaining p columns correspond to non-symmetric TS.

QAOALandscapes.transitionState — Method

transitionState(Γp::Vector{Float64}, i::Int; tsType='symmetric')

Given an initial state Γp::Vector{Float64} of length 2p it creates another vector ΓTs of size 2p+2 such that the i-th γ component of ΓTs is 0 and the i-th (i-1-th) β component of ΓTs is zero if tsType='symmetric' (tsType='non_symmetric') while all the other components are the same as Γp

Keyword arguments

tsType='symmetric' Only strings values 'symmetric' and 'non_symmetric' are accepted